feat: Add "Add Competence" button to CompetencesTable
This commit adds a button to the CompetencesTable component that allows users to add a new competence. The button is displayed at the bottom of the table and is styled using the Button component from the UI library. Clicking on the button will navigate the user to the "/competences/add" page. This addition enhances the user experience by providing an intuitive way to add competences directly from the competences table.