feat: Update IDatabase and related classes to use Omit<Cours, "_id" | "createdAt">
This commit updates the IDatabase interface and its related classes to use the Omit<Cours, "_id" | "createdAt"> type for the "cours" parameter. This change ensures that the "_id" and "createdAt" properties are excluded from the "cours" object when performing operations such as adding a new cours. By doing so, it improves the consistency and integrity of the data stored in the database.