feat: Add useRouter hook for navigation
This commit adds the useRouter hook from the "next/navigation" package to the "useSaveCoursMetadata" service. The useRouter hook allows for programmatic navigation to the newly created cours page after saving the cours metadata. This addition enhances the user experience by automatically redirecting the user to the appropriate page after completing the save operation.