refactor: Add HeadingMenuBar component for managing headings in the editor
This commit adds the HeadingMenuBar component to the cours/presentation/components directory. The component provides a dropdown menu with options for setting different heading levels in the editor. It uses the @tiptap/react and @tiptap/extension-heading libraries to interact with the editor and set the appropriate heading level. This addition improves the user experience by allowing easy management of headings within the editor.