refactor: Update CoursRepository to use Partial<Omit<Cours, "_id" | "createdAt">> for cours parameter
This commit updates the CoursRepository to use the Partial<Omit<Cours, "_id" | "createdAt">> type for the cours parameter in the addCours and updateCours methods. This change allows users to update various properties of a course while excluding the _id and createdAt fields. It improves the maintainability and flexibility of the codebase by providing a more specific type for the cours parameter.