feat: Add CoursSaveButton component
This commit adds the CoursSaveButton component to the cours-sequence feature. The CoursSaveButton component is responsible for rendering a button that allows users to save the changes made to a cours. It utilizes the useUpdateCoursBody hook to update the cours body in the database. This addition enhances the functionality of the feature by providing a convenient way for users to save their changes.