feat: Add addSequence method to CoursRepository
This commit adds the addSequence method to the CoursRepository in the cours-sequence feature. The addSequence method allows users to add a new sequence to the database with the provided metadata. It takes in the userId and sequence object as parameters and utilizes the _db.addSequence method to perform the database operation. This addition enhances the functionality of the feature by providing a convenient way for users to add sequences.