feat: Add AddSequenceView component
This commit adds the AddSequenceView component to the cours-sequence feature. The AddSequenceView component is responsible for rendering a view that allows users to add a new sequence. It provides a simple UI with the text "AddSequenceView". This addition enhances the functionality of the feature by providing a dedicated view for adding sequences.