refactor: Update CoursSequenceView to display sequence name and add "Add a cours" button
This commit refactors the CoursSequenceView component in the cours-sequence feature. It updates the code to display the sequence name as a heading and adds a button labeled "Add a cours" that allows users to add a new cours to the sequence. This change improves the user experience by providing clear visual cues and an intuitive way to add cours to a sequence.