feat: Add imageUrl property to CoursMetadata interface
This commit adds the "imageUrl" property to the "CoursMetadata" interface in the "useSaveCoursMetadata" service. The "imageUrl" property allows for the storage and retrieval of image URLs associated with each course. This change enhances the functionality and visual presentation of the application by enabling the display of images related to courses.