feat: Add button for adding sequence to SequencesListView
This commit adds a button for adding a new sequence to the SequencesListView component. The button is rendered at the bottom of the grid of sequence cards and is styled as an outline button. When clicked, it redirects the user to the "/sequences/add" page. This addition enhances the user experience by providing an intuitive way to add new sequences directly from the sequences list view.