feat: Add getAllCoursFromSequence method to retrieve all cours in a sequence
This commit adds the getAllCoursFromSequence method to the ConvexDatabase class and the CoursRepository class. The method takes in a userId and sequenceId as parameters and retrieves all cours associated with the specified sequence. It uses the getAllCoursInSequence query to fetch the cours from the database. This addition enhances the functionality of the application by allowing users to retrieve and display all cours in a sequence.