feat: Add "Complements" navigation item to side navigation
Add a new navigation item titled "Complements" to the side navigation menu in the side-nav.tsx file. This change includes the addition of the "Complements" item with an icon and a link to the "/complements/add" page. It enhances the functionality of the codebase by providing support for navigating to the "Complements" section. This commit aligns with recent user commits and expands the navigation options in the application.