feat: Add @radix-ui/react-switch package
Add the @radix-ui/react-switch package to the project dependencies. This change includes updating the package.json and pnpm-lock.yaml files to include the "@radix-ui/react-switch" package with version "^1.0.3". The @radix-ui/react-switch package provides a switch component for toggling between two states. This enhancement expands the functionality of the codebase by introducing a new UI component for user interaction.